Increased timeout to 1 hour from 30 min
Added _TURN_OFFFANS macro to make it easier to end prints

Set max extrude to 100mm for easier color change and extrude testing

Added bed fans support (need to tune how they come on - maybe thermostatic - as they
can overrun the bed heater when things are cold and klipper shutwdown thinking the bed
temp is not rising at the rate it was expecting.  This is less of an issue when the bed
is fully up to temp - 100% bed fans works then.

Tuned z probe offset so most prints can start with 0 Z adjustment on smooth PEI